+++ support/header.js
@@ -1,6 +1,0 @@
-/*!
-
-image-blob-reduce
-https://github.com/nodeca/image-blob-reduce
-
-*/